move edit profile form to seperate component, handle undefined data
