+++ components/EditProfileForm/index.ts
@@ -1,0 +1,2 @@
+export * from './EditProfileForm';
+export * from './EditProfileForm.interface';
